internal/generator/templates/template_test.go
package templates

import (
    "image"
    "testing"
    "path/filepath"
    "os"
    "runtime"

    "github.com/ControlYourPotatoes/card-generator/internal/card"
    "github.com/ControlYourPotatoes/card-generator/internal/generator/layout"
)

// getTemplateDir returns the path to the templates directory relative to the test file
func getTemplateDirTest(t *testing.T) string {
    // Get the directory containing the current test file
    _, filename, _, ok := runtime.Caller(0)
    if !ok {
        t.Fatal("Could not get current file path")
    }
    
    // Log the test file location for debugging
    t.Logf("Test file location: %s", filename)
    
    // Get the directory containing the test file
    testDir := filepath.Dir(filename)
    t.Logf("Test directory: %s", testDir)
    
    // Template directory should be at images/
    templateDir := filepath.Join(testDir, "images")
    
    // Check if directory exists and log the result
    if _, err := os.Stat(templateDir); os.IsNotExist(err) {
        t.Logf("Warning: Template directory not found at: %s", templateDir)
        
        // Try to list parent directory contents for debugging
        parentDir := filepath.Dir(testDir)
        entries, err := os.ReadDir(parentDir)
        if err == nil {
            t.Log("Contents of parent directory:")
            for _, entry := range entries {
                t.Logf("  - %s", entry.Name())
            }
        }
    } else {
        // List contents of template directory
        entries, err := os.ReadDir(templateDir)
        if err == nil {
            t.Log("Contents of template directory:")
            for _, entry := range entries {
                t.Logf("  - %s", entry.Name())
            }
        }
    }
    
    return templateDir
}

func TestTemplateCreation(t *testing.T) {
    templatesPath := getTemplateDir(t)
    
    tests := []struct {
        name      string
        cardType  card.CardType
        imageName string
    }{
        {"Creature", card.TypeCreature, "BaseCreature.png"},
        {"Artifact", card.TypeArtifact, "BaseArtifact.png"},
        {"Spell", card.TypeSpell, "BaseSpell.png"},
        {"Incantation", card.TypeIncantation, "BaseIncantation.png"},
        {"Anthem", card.TypeAnthem, "BaseAnthem.png"},
    }

    for _, tt := range tests {
        t.Run(tt.name, func(t *testing.T) {
            imagePath := filepath.Join(templatesPath, tt.imageName)
            t.Logf("Looking for template image at: %s", imagePath)
            
            // Test if file exists before trying to load it
            if _, err := os.Stat(imagePath); os.IsNotExist(err) {
                t.Fatalf("Template image not found at: %s", imagePath)
            }
            
            template, err := NewTemplate(tt.cardType)
            if err != nil {
                t.Fatalf("Failed to create template: %v", err)
            }
            if template == nil {
                t.Error("Template should not be nil")
            }

            frame, err := LoadFrame(imagePath)
            if err != nil {
                t.Fatalf("Failed to load frame from %s: %v", imagePath, err)
            }

            bounds := frame.Bounds()
            if bounds.Empty() {
                t.Error("Frame bounds should not be empty")
            }

            // Log successful frame load
            t.Logf("Successfully loaded frame for %s with dimensions %dx%d",
                tt.name, bounds.Dx(), bounds.Dy())
        })
    }
}

func TestFrameLoading(t *testing.T) {
    tests := []struct {
        name     string
        cardType card.CardType
        cardData *card.CardData
        wantFile string
    }{
        {
            name:     "Basic Creature",
            cardType: card.TypeCreature,
            cardData: &card.CardData{
                Type:    card.TypeCreature,
                Name:    "Test Creature",
                Cost:    1,
                Effect:  "Test effect",
                Attack:  1,
                Defense: 1,
            },
            wantFile: "BaseCreature.png",
        },
        {
            name:     "Basic Artifact",
            cardType: card.TypeArtifact,
            cardData: &card.CardData{
                Type:   card.TypeArtifact,
                Name:   "Test Artifact",
                Cost:   1,
                Effect: "Test effect",
            },
            wantFile: "BaseArtifact.png",
        },
    }

    for _, tt := range tests {
        t.Run(tt.name, func(t *testing.T) {
            template, err := NewTemplate(tt.cardType)
            if err != nil {
                t.Fatalf("Failed to create template: %v", err)
            }

            frame, err := template.GetFrame(tt.cardData)
            if err != nil {
                t.Fatalf("Failed to get frame: %v", err)
            }
            if frame == nil {
                t.Error("Frame should not be nil")
            }

            // Verify the frame dimensions
            bounds := frame.Bounds()
            if bounds.Empty() {
                t.Error("Frame bounds should not be empty")
            }
        })
    }
}

func TestTextBounds(t *testing.T) {
    template, err := NewTemplate(card.TypeCreature)
    if err != nil {
        t.Fatalf("Failed to create template: %v", err)
    }

    cardData := &card.CardData{
        Type:    card.TypeCreature,
        Name:    "Test Creature",
        Cost:    1,
        Effect:  "Test effect",
        Attack:  1,
        Defense: 1,
    }

    bounds := template.GetTextBounds(cardData)
    if bounds == nil {
        t.Fatal("Bounds should not be nil")
    }

    // Expected bounds for a creature card
    expectedBounds := &layout.TextBounds{
        Name: layout.TextConfig{
            Bounds:    image.Rect(125, 90, 1375, 170),  // Updated to match actual values
            FontSize:  72,
            Alignment: "center",
        },
        Effect: layout.TextConfig{
            Bounds:    image.Rect(160, 1250, 1340, 1750),
            FontSize:  48,
            Alignment: "left",
        },
        Stats: &layout.StatsConfig{
            Left: layout.TextConfig{
                Bounds:    image.Rect(130, 1820, 230, 1900),
                FontSize:  72,
                Alignment: "center",
            },
            Right: layout.TextConfig{
                Bounds:    image.Rect(1270, 1820, 1370, 1900),
                FontSize:  72,
                Alignment: "center",
            },
        },
    }

    // Compare bounds
    if bounds.Effect.Bounds != expectedBounds.Effect.Bounds {
        t.Errorf("Effect bounds = %v, want %v", bounds.Effect.Bounds, expectedBounds.Effect.Bounds)
    }
    if bounds.Name.Bounds != expectedBounds.Name.Bounds {
        t.Errorf("Name bounds = %v, want %v", bounds.Name.Bounds, expectedBounds.Name.Bounds)
    }
    if bounds.Stats == nil {
        t.Error("Stats bounds should not be nil")
    } else if bounds.Stats.Left.Bounds != expectedBounds.Stats.Left.Bounds {
        t.Errorf("Stats left bounds = %v, want %v", bounds.Stats.Left.Bounds, expectedBounds.Stats.Left.Bounds)
    }
}

func TestCreatureStats(t *testing.T) {
    template, err := NewTemplate(card.TypeCreature)
    if err != nil {
        t.Fatalf("Failed to create template: %v", err)
    }

    cardData := &card.CardData{
        Type:    card.TypeCreature,
        Name:    "Test Creature",
        Attack:  1,
        Defense: 1,
    }

    bounds := template.GetTextBounds(cardData)
    if bounds.Stats == nil {
        t.Error("Creature template should include stat bounds")
        return
    }

    // Check stat positions
    if bounds.Stats.Left.Bounds.Empty() {
        t.Error("Left stat bounds should not be empty")
    }
    if bounds.Stats.Right.Bounds.Empty() {
        t.Error("Right stat bounds should not be empty")
    }

    // Verify stat positions are within card bounds
    maxWidth := 1500  // Card width
    maxHeight := 2100 // Card height

    if bounds.Stats.Left.Bounds.Max.X > maxWidth || 
       bounds.Stats.Left.Bounds.Max.Y > maxHeight {
        t.Error("Left stat bounds exceed card dimensions")
    }

    if bounds.Stats.Right.Bounds.Max.X > maxWidth || 
       bounds.Stats.Right.Bounds.Max.Y > maxHeight {
        t.Error("Right stat bounds exceed card dimensions")
    }
}